refactor: 클래스 분리
- 로또에서 사용할 부분만 Tokenizer로 분리
   - 로또에서 필요없는 기능은 private 메소드로 전환, 테스트 제거
- ErrorMessage 추가
